feat: 홈 페이지 마크업 - 루트로 접근 시, 로그인 상태에 따라 home으로 routing 처리 - 로그인 페이지 > 헤더 미노출 처리
