+++ src/components/layout/Navbar.tsx
@@ -1,0 +1,18 @@
+'use client';
+import React, { useEffect, useState } from 'react';
+import { IcWrite } from '@assets/icons';
+import Link from 'next/link';
+import { IconButton } from '@components/elements/buttons/IconButton';
+import RoundButton from '@components/elements/buttons/RoundButton';
+import Avatar from '@components/elements/avatars/Avatar';
+
+type NavbarProps = {
+	isLogin: boolean;
+	onClickAvatar: () => void;
+};
+
+const Navbar = () => {
+	return <div></div>;
+};
+
+export default Navbar;
